+++ dev_utils/debug_dashboard/sec_utils.py
@@ -53,12 +53,11 @@ def get_emoji_chain(cls: type):
         emoji = {
             se.TextElement: "📝",
             se.TitleElement: "🏷️",
-            se.RootSectionElement: "📚",
+            se.TopLevelSectionStartMarker: "📚",
             se.TableElement: "📊",
             se.ImageElement: "🖼️",
             se.UndeterminedElement: "🛸",
             se.IrrelevantElement: "🚮",
-            se.RootSectionSeparatorElement: "⏸️",
             se.EmptyElement: "0️⃣",
             se.BulletpointTextElement: "✏️",
             se.FootnoteTextElement: "↙️",
@@ -124,4 +123,3 @@ def get_accession_number_from_url(url):
         msg = "Input string must be 18 characters long"
         raise ValueError(msg)
     return s[:10] + "-" + s[10:12] + "-" + s[12:]
-