+++ sec_parser/processing_steps/__init__.py
@@ -5,18 +5,17 @@ subpackage. These steps carry out specific tasks such as
 section identification, title parsing, and text extraction, etc.
 """
 
-from sec_parser.processing_steps.abstract_processing_step import AbstractTransformStep
+from sec_parser.processing_steps.abstract_processing_step import AbstractProcessingStep
 from sec_parser.processing_steps.footnote_and_bulletpoint_parsing_step import (
     FootnoteAndBulletpointParsingStep,
 )
 from sec_parser.processing_steps.image_parsing_step import ImageParsingStep
-from sec_parser.processing_steps.root_section_parsing_step import RootSectionParsingStep
 from sec_parser.processing_steps.table_parsing_step import TableParsingStep
 from sec_parser.processing_steps.text_parsing_step import TextParsingStep
 from sec_parser.processing_steps.title_parsing_step import TitleParsingStep
 
 __all__ = [
-    "AbstractTransformStep",
+    "AbstractProcessingStep",
     "RootSectionParsingStep",
     "TextParsingStep",
     "TitleParsingStep",