+++ sec_parser/processing_steps/abstract_elementwise_processing_step.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from typing import Callable
 
 from sec_parser.exceptions import SecParserValueError
-from sec_parser.processing_steps.abstract_processing_step import AbstractTransformStep
+from sec_parser.processing_steps.abstract_processing_step import AbstractProcessingStep
 from sec_parser.semantic_elements.abstract_semantic_element import (
     AbstractSemanticElement,
 )
@@ -21,30 +21,28 @@ class ElementwiseProcessingContext:
     """
 
     # The is_root variable informs the processing step whether the given
-    # semantic element has an HTML tag that is at the root level of the
+    # semantic element wraps an HTML tag that is at the top level of the
     # HTML document.
     is_root_element: bool
 
 
-class AbstractElementwiseTransformStep(AbstractTransformStep):
+class AbstractElementwiseProcessStep(AbstractProcessingStep):
     """
-    `AbstractElementwiseTransformStep` class applies transformations
-    to a list of semantic and container elements. Transformations are
-    applied recursively, element by element. The class can also be
-    used to iterate over all elements without applying transformations.
+    `AbstractElementwiseTransformStep` class is used to iterate over
+    all Semantic Elements with or without applying transformations.
     """
 
     def __init__(
         self,
         *,
-        process_only: set[type[AbstractSemanticElement]] | None = None,
-        except_dont_process: set[type[AbstractSemanticElement]] | None = None,
+        types_to_process: set[type[AbstractSemanticElement]] | None = None,
+        types_to_exclude: set[type[AbstractSemanticElement]] | None = None,
     ) -> None:
         super().__init__()
-        self._processed_types = process_only or set()
-        self._then_excluded_types = except_dont_process or set()
-        if self._processed_types & self._then_excluded_types:
-            msg = "Processed types and ignored types should not overlap."
+        self._types_to_process = types_to_process or set()
+        self._types_to_exclude = types_to_exclude or set()
+        if self._types_to_process & self._types_to_exclude:
+            msg = "Processed types and excluded types should not overlap."
             raise SecParserValueError(msg)
 
     def _process(
@@ -58,14 +56,14 @@ class AbstractElementwiseTransformStep(AbstractTransformStep):
         )
 
         for i, input_element in enumerate(elements):
-            if self._processed_types and not any(
-                isinstance(input_element, t) for t in self._processed_types
+            if self._types_to_process and not any(
+                isinstance(input_element, t) for t in self._types_to_process
             ):
                 continue
-            if any(isinstance(input_element, t) for t in self._then_excluded_types):
+            if any(isinstance(input_element, t) for t in self._types_to_exclude):
                 continue
 
-            element = self._transform_element(input_element, context)
+            element = self._process_element(input_element, context)
 
             if element.inner_elements:
                 child_context = ElementwiseProcessingContext(
@@ -81,13 +79,13 @@ class AbstractElementwiseTransformStep(AbstractTransformStep):
         return elements
 
     @abstractmethod
-    def _transform_element(
+    def _process_element(
         self,
         element: AbstractSemanticElement,
         context: ElementwiseProcessingContext,
     ) -> AbstractSemanticElement:
         """
-        `transform_element` method is responsible for transforming a
+        `_process_element` method is responsible for transforming a
         single semantic element into another.
 
         It can also be utilized to simply iterate over all