+++ sec_parser/semantic_elements/semantic_elements.py
@@ -15,11 +15,12 @@ class UndeterminedElement(AbstractSemanticElement):
     """
 
 
-class RootSectionElement(AbstractSemanticElement):
+class TopLevelSectionStartMarker(AbstractSemanticElement):
     """
-    The RootSectionElement class represents the top-level section of a document.
-    For instance, in SEC 10-Q reports, a RootSection could be "Part I, Item 3.
-    Quantitative and Qualitative Disclosures About Market Risk.".
+    The TopLevelSectionStartMarker class represents the beginning of a top-level
+    section of a document. For instance, in SEC 10-Q reports, a
+    top-level section could be "Part I, Item 3. Quantitative and Qualitative
+    Disclosures About Market Risk.".
     """
 
 
@@ -33,7 +34,7 @@ class IrrelevantElement(AbstractSemanticElement):
     """
 
 
-class EmptyElement(IrrelevantElement):
+class EmptyElement(AbstractSemanticElement):
     """
     The EmptyElement class represents an HTML element that does not contain any content.
     It is a subclass of the IrrelevantElement class and is used to identify and handle
@@ -41,19 +42,7 @@ class EmptyElement(IrrelevantElement):
     """
 
 
-class RootSectionSeparatorElement(IrrelevantElement):
-    """
-    The RootSectionSeparatorElement class represents a tag <document-root-section>,
-    or any tag that contains it.
-
-    We're currently using *sec-api.io* to handle the removal of the
-    title 10-Q page and to download 10-Q Section HTML files. The sections
-    are then joined by inserting a <document-root-section> separator. In the
-    future, we aim to download these HTML files directly from the SEC EDGAR.
-    """
-
-
-class TitleElement(AbstractLevelElement):
+class TitleElement(AbstractLevelElement, AbstractSemanticElement):
     """
     The TitleElement class represents the title of a paragraph or other content object.
     It serves as a semantic marker, providing context and structure to the document.